requirements.txt
streamlit==1.45.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-huggingface==0.2.0
langchain-together==0.3.0
# google-generativeai==0.8.5
agno==1.4.6
protobuf==3.20.3
duckduckgo-search==8.0.1
# chromadb==1.0.8
chromadb>=0.4.13,<0.5.0
beautifulsoup4==4.13.4
pymupdf==1.25.5
numpy<2.0.0
pysqlite3-binary
# langchain-google-genai==2.1.4